Registering conflict of loaded modules to ensure they are satisfied
When defining a conflict with module "b" in module "a", this conflict is
not satisfied when loading module "b" unless this module defines a
conflict with "a". Since it happens that modules "a" and "b" are not
managed by the same people, it seems interesting to guaranty that
conflicts defined by loaded modules are always satisfied.

To do so, this patch introduces the registration of the conflict
declared by loaded modules in _LMCONFLICT_ environment variable. Checks
have been added to cmdModuleLoad procedure to forbid the load of
modules targeted by a registered conflict. Since modulefile may directly
call module load, an is-loading procedure has been introduced to also
ensure that current module will not be loaded if it defines a conflict
toward a module which is currently loading.

Original load control behavior is preserved by default. New behavior
can be enabled by setting to 0 the g_orig_load_control variable in
modulecmd.tcl script.