Registering prereq of loaded modules to ensure they are satisfied
When defining a prereq with module "b" in module "a", this prereq is
not satisfied when module "b" is unloaded whereas module "a" is still
loaded. It seems interesting to guaranty that prereq modules cannot be
unloaded if modules requiring them are still loaded.

To do so, this patch introduces the registration of the prereq declared
by loaded modules in _LMPREREQ_ environment variable. Checks have been
added to cmdModuleUnload procedure to forbid the unload of modules
targeted by a registered prereq.

In case of a switch action, all modules declaring a prereq on the switched
module are reloaded. cmdModuleSwitch procedure first unloads the modules
declaring prereq, then proceed switch as usual and after that loads
again the modules declaring prereq.

Similarly to switch action, modules declaring an optional prereq on a
module are reloaded when this module is loaded or unloaded.
cmdModuleLoad and cmdModuleUnload procedures reload the modules
declaring prereq after their main load/unload process to avoid reloading
modules whereas prereq-module fails to load/unload.

Original load control behavior is preserved by default. New behavior
can be enabled by setting to 0 the g_orig_load_control variable in
modulecmd.tcl script.